Use single definition for cappedCanvasSize()
